Moved Metadata to general app and changed points accordingly
