+++ wazimap_ng/general/serializers.py
@@ -1,0 +1,10 @@
+from . import models
+from wazimap_ng.datasets.serializers import LicenceSerializer
+
+
+class MetaDataSerializer(serializers.ModelSerializer):
+    licence = LicenceSerializer(read_only=True)
+
+    class Meta:
+        model = models.MetaData
+        fields = ('source', 'description', 'licence',)
