+++ wazimap_ng/points/serializers.py
@@ -56,10 +56,3 @@ class ProfileCategorySerializer(serializers.ModelSerializer):
         model = models.ProfileCategory
         fields = ('id', 'label', 'description', 'theme', 'theme_id', 'theme_icon', 'subtheme', 'subtheme_id')
         #fields = ('id', 'label', 'description', 'theme', 'theme_id', 'theme_icon', 'subtheme', 'subtheme_id', 'locations', )
-
-class MetaDataSerializer(serializers.ModelSerializer):
-    licence = LicenceSerializer(read_only=True)
-
-    class Meta:
-        model = models.MetaData
-        fields = ('source', 'description', 'licence',)
