+++ wazimap_ng/points/views.py
@@ -42,7 +42,7 @@ def theme_view(request, profile_id=None):
             js_theme["categories"].append({
                 "id": pc.category.id,
                 "name": pc.label,
-                "metadata": serializers.MetaDataSerializer(pc.metadata).data
+                "metadata": serializers.MetaDataSerializer(pc.category.metadata).data
             })
             
         js.append(js_theme)
